Add files via upload
Developed a proof-of-concept app for a "your business" using SQL, an Excel front end, and VBA. This system can be used in hotels or some similar industries because it contains the information of customers, products, and sales. In Access, the customers’ names, email addresses, countries can be founded in the table. The products’ type, unit price, stocks can also be observed in its table, and if other industries would like to use this system, you only need to change “RoomTypes” to your product’s types. For the sales table, the order’s transaction date, quantity, and status can be observed easily. Also, the queries are useful to observe some issues, such as the revenue of a month, the most popular product in a month, and which countries’ people order the most. In Excel, we can import the database from Access, and we can also export the new information back to Access. Furthermore, in the system, we can search the details from entering the customer’s name or add a new customer to the customer list.